Add support for search pagination (#15)
Always add the search pagination control to each search request to
ensure that large data sets can be fully enumerated. The presence of the
control will be ignored by servers that do not support the control code
so a warning is displayed for those scenarios. Also fixes the control
code response parser to properly decode the incoming value that failed
in previous scenarios.