Process referrals and add integration environment (#16)
Will ensure any search result that contains a referral is emitted as an
error record to the caller. This ensures that the referral isn't lost
and that it provides enough detail for the caller to then resend the
request to the correct server. This replicates how the AD cmdlets work
when a referral is returned.

Also adds a very basic structure for an integration environment to test
out scernarios like referrals. This will hopefully be expanded in the
future to run automated tests and flesh out the environment a bit more.