Expand Get-OpenADWhoami Output
Expand the output of `Get-OpenADWhoami` to include a parsed username
response without the leading `u:` and also include other session
information like the domain controller DNS name under
`DomainController`, the LDAP uri used for the connection under `Uri`,
and the authentication method used for the connection under
`Authentication`.

Also include the `DomainController` property on the OpenADSession class
which is based on the `dnsHostName` attribute on the rootDSE queried
when the session was created.