Fix default parameter sets for Get-OpenAD* cmdlets
Fixes the default parameter set used for the Get-OpenAD* cmdlets so that
the default LDAPFilter that selects all of that type is always applied
unless an explicit filter was specified or an identity was specified.
This means setting a search base will no longer prompt for a filter to
use.