Upper case the output LDAP attribute properties (#41)
Make the output note properties for each LDAP attribute start with an
upper cased letter. This makes the object fit more in line with other
.NET objects and makes things a bit nicer to read when it comes to
generating reports for these objects. Because PowerShell is case
insensitive when getting properties this will be a backwards compatible
change. Users can still select the property using the LDAP attribute
casing.