Add better property validation
The Get-OpenAD* cmdlets will attempt to validate that the requested
properties are valid for the object type being requested. It does this
by consulting the list of attributes defined for the objectClass being
queries as well as any auxiliary class that is associated with it.
Attempting to use an invalid property will cause an error.

Properties that are explicitly requested but are not set will still be
present on the request object but with a `$null` value.