Fix get operation without property metadata (#49)
Fix get operations where the session does not have any schema metadata
to check the properties with. This is mostly a rare scenario, like
unauthenticated binds, where no schema metadata could be retrieved as
the unauthenticated bind does not have permission to do so. This fact
will be seen in the search response and is preferable to a "not a valid
attribute message" that was previously shown.

Also bumps the version to 0.1.0 actual as I feel it is ready.