Refactor Kerberos and make state Runspace
Refactors the Kerberos code to new standards used in some of my
projects. Also includes the Kerberos error when failing to find the
principal for better debugging purposes.

Sets the GlobalState to be Runspace specific rather than process wide.
This will allow the module to be used in separate runspaces like with
ForEach-Object -Parallel or other Runspace based APIs.